changes based on comments #108
