+++ templates/back/profil/_modal_profil_password.html.twig
@@ -4,7 +4,7 @@
             <div class="fr-col-12 fr-col-md-8">
                 <div class="fr-modal__body">
                     <form action="{{ path('back_profil_edit_password') }}" name="login-creation-mdp-form"
-                        id="profil_edit_password_form" method="POST" {# enctype="multipart/form-data" #}>
+                        id="profil_edit_password_form" method="POST">
                         <div class="fr-modal__header">
                         <button type="button" class="fr-btn--close fr-btn" aria-controls="fr-modal-profil-edit-password">Fermer</button>
                         </div>
