+++ tests/Unit/Twig/AppExtensionTest.php
@@ -4,15 +4,8 @@ declare(strict_types=1);
 
 namespace App\Tests\Unit\Twig;
 
-use App\Entity\Territory;
-use App\Entity\User;
-use App\Service\TimezoneProvider;
 use App\Twig\AppExtension;
-use League\Flysystem\FilesystemOperator;
-use PHPUnit\Framework\MockObject\MockObject;
 use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
-use Symfony\Bundle\SecurityBundle\Security;
-use Symfony\Component\DependencyInjection\ParameterBag\ParameterBagInterface;
 
 class AppExtensionTest extends WebTestCase
 {
@@ -97,54 +90,5 @@ class AppExtensionTest extends WebTestCase
             self::TWIG_DATE_FORMAT_DEFAULT,
             'UTC',
         ];
-    }
-
-    public function testUserAvatarOrPlaceHolder(): void
-    {
-        self::bootKernel();
-        $container = static::getContainer();
-        $appExtension = $container->get(AppExtension::class);
-
-        $user = (new User())
-        ->setRoles([User::ROLES['Super Admin']]);
-        $outputSpan = $appExtension->userAvatarOrPlaceHolder($user);
-        $this->assertEquals('<span class="avatar-placeholder avatar-74">SA</span>', $outputSpan);
-
-        $outputSpan = $appExtension->userAvatarOrPlaceHolder($user, 80);
-        $this->assertEquals('<span class="avatar-placeholder avatar-80">SA</span>', $outputSpan);
-
-        $user->setRoles([User::ROLES['Administrateur']]);
-        $user->setTerritory((new Territory())->setZip('44'));
-        $outputSpan = $appExtension->userAvatarOrPlaceHolder($user);
-        $this->assertEquals('<span class="avatar-placeholder avatar-74">44</span>', $outputSpan);
-
-        /** @var Security $security */
-        $security = $this->createMock(Security::class);
-        /** @var TimezoneProvider $timeZoneProvider */
-        $timeZoneProvider = new TimezoneProvider($security);
-        /** @var ParameterBagInterface $parameterBag */
-        $parameterBag = $this->createMock(ParameterBagInterface::class);
-        /** @var FilesystemOperator|MockObject $fileStorage */
-        $fileStorage = $this->createMock(FilesystemOperator::class);
-        $fileStorage->expects($this->exactly(2))
-            ->method('fileExists')
-            ->willReturn(true);
-
-        $appExtension = new AppExtension(
-            $timeZoneProvider,
-            $parameterBag,
-            $fileStorage,
-        );
-
-        $user->setAvatarFilename(__DIR__.'/../../files/sample.jpg');
-        $outputSpan = $appExtension->userAvatarOrPlaceHolder($user);
-
-        $this->assertStringContainsString('<img src="data:image/jpg;base64,', $outputSpan);
-        $this->assertStringContainsString('alt="Avatar de l\'utilisateur" class="avatar-74">', $outputSpan);
-
-        $outputSpan = $appExtension->userAvatarOrPlaceHolder($user, 100);
-
-        $this->assertStringContainsString('<img src="data:image/jpg;base64,', $outputSpan);
-        $this->assertStringContainsString('alt="Avatar de l\'utilisateur" class="avatar-100">', $outputSpan);
     }
 }
